cleaning up vulcan root UI
